Massive commit of clean up using static code analysis
Use IntelliJ Code Cleanup

+++ engine/src/main/java/com/arcadedb/engine/MurmurHash.java
@@ -167,7 +167,7 @@ public final class MurmurHash {
     case 2:
       h ^= (long) (data[(length & ~7) + 1] & 0xff) << 8;
     case 1:
-      h ^= (long) (data[length & ~7] & 0xff);
+      h ^= data[length & ~7] & 0xff;
       h *= m;
     }
